# docs/swagger/swagger.yaml
openapi: 3.0.4
info:
  title: Flood Rescue & Relief System API
  description: |-
    API documentation for Flood Rescue & Relief Management System.

    This system manages rescue and relief operations during flood emergencies, including:
    - User authentication and authorization
    - Emergency rescue/relief requests
    - Mission assignment and tracking
    - Resource and inventory management
    - Real-time team position tracking
    - Notifications and reporting

    **Roles:**
    - Citizen: Can submit requests and view their status
    - Rescue Team: Receives and executes missions
    - Rescue Coordinator: Manages requests and assigns missions
    - Manager: Manages resources and inventory
    - Admin: Full system access

    **Production Server:** https://flood-rescue.onrender.com
  contact:
    email: duylase184704@fpt.edu.vn
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  version: 1.0.0
servers:
  - url: https://flood-rescue.onrender.com
    description: Production server
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/missions:
    post:
      tags:
        - Missions
      summary: Create mission
      description: "Create a new mission (initial status = DRAFT). Request/team planning is done via /{id}/requests and /{id}/teams. (Roles: Rescue Coordinator, Admin)"
      operationId: createMission
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMissionInput"
      responses:
        "201":
          description: Mission created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mission"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Forbidden - Coordinator/Admin only
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    get:
      tags:
        - Missions
      summary: List missions
      description: "Get list of missions with pagination and filtering. (Roles: Rescue Coordinator, Admin)"
      operationId: listMissions
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          description: Filter by mission status
          required: false
          schema:
            type: string
            enum: [DRAFT, PLANNED, IN_PROGRESS, PAUSED, PARTIAL, COMPLETED, ABORTED]
        - name: type
          in: query
          description: Filter by mission type
          required: false
          schema:
            type: string
            enum: [RESCUE, RELIEF]
        - name: code
          in: query
          description: Filter by mission code (partial match)
          required: false
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: Missions retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Mission"
                  meta:
                    $ref: "#/components/schemas/Pagination"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /api/missions/{id}:
    get:
      tags:
        - Missions
      summary: Get mission detail
      description: Get detailed information about a specific mission
      operationId: getMissionDetail
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: Mission ID
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Mission detail retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mission"
        "404":
          description: Mission not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    patch:
      tags:
        - Missions
      summary: Update mission
      description: "Update mission name, description or priority. Status cannot be updated here — use /pause, /resume, /abort. (Roles: Rescue Coordinator, Admin)"
      operationId: updateMission
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: Mission ID
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMissionInput"
      responses:
        "200":
          description: Mission updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mission"
        "400":
          description: Invalid input or direct status update attempted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Mission not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      tags:
        - Missions
      summary: Delete mission
      description: "Delete a mission. Cannot delete if there are active timelines. (Roles: Rescue Coordinator, Admin)"
      operationId: deleteMission
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: Mission ID
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Mission deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Mission deleted successfully
        "400":
          description: Cannot delete mission with active timelines
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Mission not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /api/missions/{id}/requests:
    get:
      tags:
        - Missions
      summary: List requests in a mission
      description: "Get all MissionRequest records associated with the mission. (Roles: Rescue Coordinator, Admin, Rescue Team)"
      operationId: getMissionRequests
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: Mission ID
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Mission requests retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/MissionRequest"
        "404":
          description: Mission not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    post:
      tags:
        - Missions
      summary: Add requests to mission
      description: "Add one or many request IDs to a DRAFT mission and create MissionRequest records. (Roles: Rescue Coordinator, Admin)"
      operationId: addRequestsToMission
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: Mission ID
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MissionAddRequestsInput"
      responses:
        "200":
          description: Requests added to mission successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/MissionRequest"
        "400":
          description: Mission is not DRAFT or request status is invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Mission or Request not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /api/missions/{id}/requests/{requestId}:
    delete:
      tags:
        - Missions
      summary: Remove request from mission
      description: "Remove one request link from a DRAFT mission. Only PENDING mission-requests can be removed. (Roles: Rescue Coordinator, Admin)"
      operationId: removeRequestFromMission
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: Mission ID
          required: true
          schema:
            type: string
        - name: requestId
          in: path
          description: Request ID to remove from mission
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Request removed from mission successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      missionId:
                        type: string
                      requestId:
                        type: string
                      removed:
                        type: boolean
        "400":
          description: Mission is not DRAFT or mission-request status is invalid for remove
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Mission or request link not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /api/missions/{id}/teams/{teamId}:
    delete:
      tags:
        - Missions
      summary: Remove team from mission
      description: "Remove one team link from a DRAFT mission. Only PLANNED or ASSIGNED timelines can be removed. (Roles: Rescue Coordinator, Admin)"
      operationId: removeTeamFromMission
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: Mission ID
          required: true
          schema:
            type: string
        - name: teamId
          in: path
          description: Team ID to remove from mission
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Team removed from mission successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      missionId:
                        type: string
                      teamId:
                        type: string
                      removed:
                        type: boolean
        "400":
          description: Mission is not DRAFT or timeline status is invalid for remove
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Mission or team link not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /api/missions/{id}/start:
    patch:
      tags:
        - Missions
      summary: Start mission planning
      description: "Move a DRAFT mission to PLANNED and convert PLANNED timelines to ASSIGNED. Requires at least one MissionRequest and one PLANNED timeline. (Roles: Rescue Coordinator, Admin)"
      operationId: startMission
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: Mission ID
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartMissionInput"
      responses:
        "200":
          description: Mission started successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mission"
        "400":
          description: Mission is not DRAFT, has no requests, or has no PLANNED timelines
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Mission not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /api/missions/{id}/pause:
    patch:
      tags:
        - Missions
      summary: Pause mission
      description: "Pause an IN_PROGRESS mission. Active timelines continue running — pausing only prevents new team assignments. (Roles: Rescue Coordinator, Admin)"
      operationId: pauseMission
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: Mission ID
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Mission paused successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mission"
        "400":
          description: Mission is not IN_PROGRESS
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Mission not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /api/missions/{id}/resume:
    patch:
      tags:
        - Missions
      summary: Resume mission
      description: "Resume a PAUSED mission back to IN_PROGRESS. (Roles: Rescue Coordinator, Admin)"
      operationId: resumeMission
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: Mission ID
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Mission resumed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mission"
        "400":
          description: Mission is not PAUSED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Mission not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /api/missions/{id}/abort:
    patch:
      tags:
        - Missions
      summary: Abort mission
      description: |
        Force-abort a mission. Automatically cancels all active timelines (ASSIGNED, EN_ROUTE, ON_SITE).
        Cannot abort a COMPLETED or already ABORTED mission. (Roles: Rescue Coordinator, Admin)
      operationId: abortMission
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: Mission ID
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Mission aborted and all active timelines cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mission"
        "400":
          description: Mission is already closed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Mission not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
...
    CreateMissionInput:
      type: object
      required:
        - name
        - type
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 200
          example: "Rescue operation - District 9 flooding"
        type:
          type: string
          enum: [RESCUE, RELIEF]
          example: RESCUE
        description:
          type: string
          maxLength: 1000
          example: "Multiple households trapped due to flash flood"
        priority:
          type: string
          enum: [Critical, High, Normal]
          default: Normal
          example: High
    UpdateMissionInput:
      type: object
      description: At least one field required. Status cannot be updated here.
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 200
        description:
          type: string
          maxLength: 1000
        priority:
          type: string
          enum: [Critical, High, Normal]
    MissionAddRequestsInput:
      type: object
      required:
        - requestIds
      properties:
        requestIds:
          type: array
          minItems: 1
          items:
            type: string
          example:
            - "60d5ec49f1b2c8b1f8e4e1a3"
            - "60d5ec49f1b2c8b1f8e4e1a4"
        note:
          type: string
          maxLength: 500
          nullable: true
          example: "Ưu tiên cứu hộ khu vực ngập sâu"
...
    StartMissionInput:
      type: object
      properties:
        note:
          type: string
          maxLength: 500
          nullable: true
...
    MissionRequest:
      type: object
      properties:
        _id:
          type: string
          example: "65f1d2c3b4a5968776655777"
        missionId:
          type: string
          example: "65f1d2c3b4a5968776655111"
        requestId:
          type: string
          example: "65f1d2c3b4a5968776655222"
        status:
          type: string
          enum:
            - PENDING
            - IN_PROGRESS
            - PARTIAL
            - FULFILLED
            - CLOSED
            - DROPPED
        peopleNeeded:
          type: integer
          minimum: 0
        peopleRescued:
          type: integer
          minimum: 0
        peopleRemaining:
          type: integer
          minimum: 0
        fulfillmentPercent:
          type: number
          minimum: 0
          maximum: 100
        note:
          type: string
          nullable: true
        closedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Mission:
      type: object
      properties:
        _id:
          type: string
          example: "60d5ec49f1b2c8b1f8e4e1a4"
        code:
          type: string
          description: Auto-generated code (e.g. MS-250225-001)
          example: "MS-250225-001"
        name:
          type: string
          example: "Rescue operation - District 9 flooding"
        description:
          type: string
          nullable: true
        status:
          type: string
          enum:
            - DRAFT
            - PLANNED
            - IN_PROGRESS
            - PAUSED
            - PARTIAL
            - COMPLETED
            - ABORTED
          example: DRAFT
        priority:
          type: string
          enum: [Critical, High, Normal]
          example: Normal
        type:
          type: string
          enum: [RESCUE, RELIEF]
          example: RESCUE
        coordinatorId:
          type: string
          example: "60d5ec49f1b2c8b1f8e4e1a1"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 100
        totalPages:
          type: integer
          example: 5
...
    ApiError:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Lỗi khi đăng nhập"
        data:
          nullable: true
          example: null
        error:
          type: object
          properties:
            code:
              type: string
              example: "INTERNAL_ERROR"
            details:
              type: array
              nullable: true
              items:
                type: object
                properties:
                  field:
                    type: string
                    example: "completions.0.missionRequestId"
                  message:
                    type: string
                    example: "missionRequestId phải là ObjectId hợp lệ"
                  type:
                    type: string
                    example: "string.pattern.base"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT token authentication. Format: `Bearer <token>`"